+++ store/modules/stimulation/index.js
@@ -26,6 +26,11 @@ const state = () => ({
   x_axis_time_idx: 0,
   edit_mode: { status: false, protocol: "", label: "", color: "" },
   stim_status: STIM_STATUS.CALIBRATION_NEEDED,
+  hovered_pulse: {
+    idx: null,
+    indices: [],
+    color: null,
+  },
 });
 
 export default {