+++ tests/unit/components/stimulation/StimulationStudioDragAndDropPanel.spec.js
@@ -1,7 +1,7 @@
-import { mount, createLocalVue } from "@vue/test-utils";
+import { mount, createLocalVue, shallowMount } from "@vue/test-utils";
 import StimulationStudioDragAndDropPanel from "@/components/stimulation/StimulationStudioDragAndDropPanel.vue";
 import Vuex from "vuex";
-
+import Vue from "vue";
 const localVue = createLocalVue();
 localVue.use(Vuex);
 let NuxtStore;
@@ -10,21 +10,20 @@ const test_protocol_order = [
   {
     type: "Biphasic",
     src: "placeholder",
-    nest_protocols: [],
     stop_setting: "Stimulate Until Complete",
     repeat: {
       number_of_repeats: 1,
-      color: "fffff",
+      color: "#fffff",
     },
     pulse_settings: {
-      phase_one_duration: 300,
+      phase_one_duration: 3,
       phase_one_charge: 2,
-      interphase_interval: 500,
-      phase_two_duration: 100,
+      interphase_interval: 5,
+      phase_two_duration: 1,
       phase_two_charge: -5,
     },
     stim_settings: {
-      repeat_delay_interval: 0,
+      delay_interval: 0,
       total_active_duration: {
         duration: 20,
         unit: "milliseconds",
@@ -35,44 +34,16 @@ const test_protocol_order = [
     type: "Monophasic",
     src: "placeholder",
     stop_setting: "Stimulate Until Complete",
-    nested_protocols: [
-      {
-        type: "Monophasic",
-        src: "placeholder",
-        nested_protocols: [],
-        repeat: {
-          number_of_repeats: 1,
-          color: "fffff",
-        },
-        pulse_settings: {
-          phase_one_duration: 100,
-          phase_one_charge: -2,
-        },
-      },
-      {
-        type: "Monophasic",
-        src: "placeholder",
-        nested_protocols: [],
-        repeat: {
-          number_of_repeats: 1,
-          color: "fffff",
-        },
-        pulse_settings: {
-          phase_one_duration: 300,
-          phase_one_charge: 6,
-        },
-      },
-    ],
     repeat: {
       number_of_repeats: 1,
-      color: "fffff",
+      color: "#ffff1",
     },
     pulse_settings: {
-      phase_one_duration: 300,
+      phase_one_duration: 3,
       phase_one_charge: 2,
     },
     stim_settings: {
-      repeat_delay_interval: 0,
+      delay_interval: 0,
       total_active_duration: {
         duration: 0,
         unit: "milliseconds",
@@ -82,21 +53,20 @@ const test_protocol_order = [
   {
     type: "Delay",
     src: "placeholder",
-    nested_protocols: [],
     stop_setting: "Stimulate Until Complete",
     repeat: {
       number_of_repeats: 1,
-      color: "fffff",
+      color: "#ffff2",
     },
     pulse_settings: {
-      phase_one_duration: 3000,
+      phase_one_duration: 3,
       phase_one_charge: 0,
       interphase_interval: 0,
       phase_two_duration: 0,
       phase_two_charge: 0,
     },
     stim_settings: {
-      repeat_delay_interval: 0,
+      delay_interval: 0,
       total_active_duration: {
         duration: 3,
         unit: "seconds",
@@ -107,17 +77,16 @@ const test_protocol_order = [
     type: "Monophasic",
     src: "placeholder",
     stop_setting: "Stimulate Until Complete",
-    nested_protocols: [],
     repeat: {
       number_of_repeats: 1,
-      color: "fffff",
+      color: "#ffff3",
     },
     pulse_settings: {
-      phase_one_duration: 300,
+      phase_one_duration: 3,
       phase_one_charge: 2,
     },
     stim_settings: {
-      repeat_delay_interval: 0,
+      delay_interval: 0,
       total_active_duration: {
         duration: 20,
         unit: "milliseconds",
@@ -140,14 +109,14 @@ const test_protocol_list = [
       time_unit: "milliseconds",
       pulses: [
         {
-          phase_one_duration: 15000,
+          phase_one_duration: 15,
           phase_one_charge: 0,
           interphase_interval: 0,
           phase_two_duration: 0,
           phase_two_charge: 0,
         },
         {
-          phase_one_duration: 20000,
+          phase_one_duration: 20,
           phase_one_charge: 0,
           interphase_interval: 0,
           phase_two_duration: 0,
@@ -159,7 +128,6 @@ const test_protocol_list = [
           type: "Delay",
           src: "/delay-tile.png",
           stop_setting: "Stimulate Until Complete",
-          nested_protocols: [],
           repeat: { color: "d822f9", number_of_repeats: 1 },
           pulse_settings: {
             phase_one_duration: 15,
@@ -169,7 +137,7 @@ const test_protocol_list = [
             phase_two_charge: 0,
           },
           stim_settings: {
-            repeat_delay_interval: 3,
+            delay_interval: 3,
             total_active_duration: {
               duration: 20,
               unit: "milliseconds",
@@ -180,7 +148,44 @@ const test_protocol_list = [
     },
   },
 ];
+const new_mono_test_element = {
+  type: "Monophasic",
+  repeat: { color: "", number_of_repeats: 0 },
+  pulse_settings: {
+    phase_one_duration: "",
+    phase_one_charge: "",
+    interphase_interval: "",
+    phase_two_duration: "",
+    phase_two_charge: "",
+  },
+  stim_settings: {
+    delay_interval: "",
+    total_active_duration: {
+      duration: "",
+      unit: "milliseconds",
+    },
+  },
+};
+const new_bi_test_element = {
+  type: "Biphasic",
+  repeat: { color: "", number_of_repeats: 0 },
+  pulse_settings: {
+    phase_one_duration: "",
+    phase_one_charge: "",
+    interphase_interval: "",
+    phase_two_duration: "",
+    phase_two_charge: "",
+  },
+  stim_settings: {
+    delay_interval: "",
+    total_active_duration: {
+      duration: "",
+      unit: "milliseconds",
+    },
+  },
+};
 
+let wrapper = null;
 describe("StimulationStudioDragAndDropPanel.vue", () => {
   beforeAll(async () => {
     const storePath = `${process.env.buildDir}/store.js`;
@@ -192,8 +197,12 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
     store.state.stimulation.protocol_list = JSON.parse(JSON.stringify(test_protocol_list));
   });
 
+  afterEach(() => {
+    wrapper.destroy();
+    jest.clearAllMocks();
+  });
   test("When mounting StimulationStudioDragAndDropPanel from the component file, Then there should be no waveforms in the new protocol container", () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
@@ -202,7 +211,7 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
   });
 
   test("When a user clicks on trash icons to delete protocol, Then the protocol order in StimulationStudioDragAndDropPanel should empty from the mutation in state", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
@@ -212,58 +221,49 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
   });
 
   test("When a user drops a waveform to the block editor, Then the corresponding modal to should pop up", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = shallowMount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
-    const modal_container = wrapper.find(".modal-container");
+
     await wrapper.vm.clone({ type: "Monophasic", src: "test" });
     await wrapper.vm.check_type({
-      added: { element: { type: "Monophasic", repeat: { color: "", number_of_repeats: 0 } } },
+      added: {
+        element: new_mono_test_element,
+        newIndex: 4,
+      },
     });
     expect(wrapper.vm.modal_type).toBe("Monophasic");
-    expect(modal_container).toBeTruthy();
 
-    await wrapper.vm.clone({ type: "Biphasic", src: "test" });
-    await wrapper.vm.check_type({
-      added: { element: { type: "Biphasic", repeat: { color: "", number_of_repeats: 0 } } },
-    });
-    expect(wrapper.vm.modal_type).toBe("Biphasic");
-    expect(modal_container).toBeTruthy();
+    const modal_container = wrapper.find(".modal-container");
+    expect(modal_container.isVisible()).toBeTruthy();
+
+    await wrapper.vm.on_modal_close("Cancel");
+    expect(wrapper.vm.modal_type).toBeNull();
   });
 
   test("When a user opens a modal to delete waveform, Then when delete is clicked, it will delete", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
-    wrapper.vm.protocol_order = test_protocol_order;
-    await wrapper.vm.open_modal_for_edit("Monophasic", 0);
-    expect(wrapper.vm.modal_type).toBe("Monophasic");
+
+    await wrapper.setData({ protocol_order: JSON.parse(JSON.stringify(test_protocol_order)) });
+    await wrapper.vm.open_modal_for_edit("Biphasic", 0);
+    expect(wrapper.vm.modal_type).toBe("Biphasic");
     expect(wrapper.vm.shift_click_img_idx).toBe(0);
     expect(wrapper.find(".modal_overlay")).toBeTruthy();
 
-    await wrapper.vm.on_modal_close("Delete");
+    const modal_buttons = wrapper.findAll(".span__button_label");
+    await modal_buttons.at(2).trigger("click");
+
     expect(wrapper.vm.protocol_order).toHaveLength(3);
     expect(wrapper.vm.modal_type).toBeNull();
-
-    await wrapper.vm.open_modal_for_edit("Biphasic", 1);
-    expect(wrapper.vm.modal_type).toBe("Biphasic");
-    expect(wrapper.vm.shift_click_img_idx).toBe(1);
-
-    await wrapper.vm.on_modal_close("Cancel");
-    expect(wrapper.vm.protocol_order).toHaveLength(3);
-    expect(wrapper.vm.modal_type).toBeNull();
-
-    await wrapper.vm.open_modal_for_edit("Biphasic", 0, 0);
-
-    await wrapper.vm.on_modal_close("Delete");
-    expect(wrapper.vm.protocol_order).toHaveLength(3);
   });
 
   test("When exiting instance, Then instance is effectively destroyed", async () => {
     const destroyed_spy = jest.spyOn(StimulationStudioDragAndDropPanel, "beforeDestroy");
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
@@ -272,7 +272,7 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
   });
 
   test("When a user selects a protocol to edit, Then the DragAndDropPanel component should get the selected pulse order and unit of time to display for edit", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
@@ -286,66 +286,52 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
     expect(wrapper.vm.protocol_order).toStrictEqual(expected_pulse_order);
   });
 
-  test("When an order changes inside of a nested loop, Then the new order should be dispatched", async () => {
-    const action_spy = jest.spyOn(store, "dispatch");
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
-      store,
-      localVue,
-    });
-    await wrapper.vm.handle_internal_repeat();
-    expect(action_spy).toHaveBeenCalledTimes(2);
-  });
-
   test("When a user adds a new waveform to the protocol editor and cancels the addition, Then the modal should only appear when it's been cloned and should remove new waveform when cancelled", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
-    wrapper.vm.protocol_order = test_protocol_order;
-    await wrapper.vm.clone({ type: "Monophasic", src: "test" });
+    await wrapper.setData({ protocol_order: JSON.parse(JSON.stringify(test_protocol_order)) });
+    expect(wrapper.vm.protocol_order).toHaveLength(4);
+
+    await wrapper.vm.clone({ type: "Biphasic", src: "test" });
     expect(wrapper.vm.cloned).toBe(true);
     await wrapper.vm.check_type({
-      added: { element: { type: "Monophasic", repeat: { color: "", number_of_repeats: 0 } }, newIndex: 3 },
+      added: {
+        element: new_bi_test_element,
+        newIndex: 4,
+      },
     });
 
-    await wrapper.vm.on_modal_close("Cancel");
-    expect(wrapper.vm.protocol_order).toHaveLength(3);
+    expect(wrapper.vm.modal_type).toBe("Biphasic");
 
-    await wrapper.vm.check_type({
-      added: { element: { type: "Biphasic", repeat: { color: "", number_of_repeats: 0 } }, newIndex: 1 },
-    });
-    expect(wrapper.vm.modal_type).toBeNull();
+    const cancel_button = wrapper.findAll(".span__button_label").at(1);
+    await cancel_button.trigger("click");
+
+    expect(wrapper.vm.protocol_order).toHaveLength(4);
   });
-
-  test("When a user creates a repeat block, Then the repeat modal should appear and save number of repeats to state", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+  test("When changes the order of waveform tiles in scrollable component, Then no modal should appear", async () => {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
-    wrapper.vm.protocol_order = test_protocol_order;
-    await wrapper.vm.handle_repeat({ added: "test" }, 1);
-    expect(wrapper.vm.repeat_delay_modal).toBe("Repeat");
-    expect(wrapper.vm.repeat_idx).toBe(1);
+    await wrapper.setData({ protocol_order: JSON.parse(JSON.stringify(test_protocol_order)) });
+    expect(wrapper.vm.protocol_order).toHaveLength(4);
 
-    await wrapper.vm.on_repeat_modal_close({ button_label: "Save", number_of_repeats: 3 });
-    await wrapper.vm.open_repeat_modal_for_edit(3, 1);
-    expect(wrapper.vm.current_repeat_delay_input).toBe(3);
+    await wrapper.vm.check_type({
+      moved: {
+        element: {},
+        newIndex: 2,
+      },
+    });
 
-    await wrapper.vm.handle_repeat({ added: "test" }, 1);
-    expect(wrapper.vm.repeat_idx).toBe(1);
-
-    await wrapper.vm.on_repeat_modal_close({ button_label: "Cancel", number_of_repeats: 3 });
-    expect(wrapper.vm.protocol_order[1].nested_protocols).toStrictEqual([]);
-
-    wrapper.vm.repeat_delay_modal = null;
-    await wrapper.vm.handle_repeat({ moved: "test" }, 1);
-    expect(wrapper.vm.repeat_delay_modal).toBeNull();
+    expect(wrapper.vm.modal_type).toBeNull();
   });
 
   test("When a user clicks save on the settings for a waveform, Then the setting should save to the corresponding index depending on if it is a new waveform or an edited", async () => {
     const test_settings = "test";
     const test_stim_settings = {
-      repeat_delay_interval: "",
+      delay_interval: "",
       total_active_duration: {
         duration: "",
         unit: "milliseconds",
@@ -381,14 +367,87 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
     expect(wrapper.vm.time_units_idx).toBe(1);
   });
 
-  test("When a user removes the last waveform from a repeat block, Then the colored border should be removed and mutations to state should be committed", async () => {
-    const wrapper = mount(StimulationStudioDragAndDropPanel, {
+  test("When a user hovers over a waveform tile, Then the pulse settings will be added to state", async () => {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
     });
-    wrapper.vm.protocol_order = test_protocol_order;
-    await wrapper.vm.handle_repeat({ removed: "test" }, 1);
-    expect(wrapper.vm.protocol_order[1].repeat.number_of_repeats).toBe(0);
+
+    expect(store.state.stimulation.hovered_pulse).toStrictEqual({
+      idx: null,
+      indices: [],
+      color: null,
+    });
+
+    await wrapper.setData({ protocol_order: JSON.parse(JSON.stringify(test_protocol_order)) });
+    await store.dispatch("stimulation/handle_protocol_order", test_protocol_order);
+    await wrapper.vm.on_pulse_enter(1);
+
+    expect(store.state.stimulation.hovered_pulse).toStrictEqual({
+      idx: 1,
+      indices: [9, 10],
+      color: "#ffff1",
+    });
+  });
+
+  test("When a user hovers over a waveform tile, but it's because the user is dragging a tile above, Then the pulse settings not be added", async () => {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
+      store,
+      localVue,
+    });
+    const default_state = {
+      idx: null,
+      indices: [],
+      color: null,
+    };
+    expect(store.state.stimulation.hovered_pulse).toStrictEqual(default_state);
+
+    await wrapper.setData({
+      protocol_order: JSON.parse(JSON.stringify(test_protocol_order)),
+      is_dragging: true,
+    });
+    await store.dispatch("stimulation/handle_protocol_order", test_protocol_order);
+    await wrapper.vm.on_pulse_enter(1);
+
+    expect(store.state.stimulation.hovered_pulse).toStrictEqual(default_state);
+  });
+
+  test("When a user leaves hover over a waveform tile, Then the pulse settings will be reset", async () => {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
+      store,
+      localVue,
+    });
+
+    await store.dispatch("stimulation/handle_protocol_order", test_protocol_order);
+    await wrapper.vm.on_pulse_enter(1);
+    expect(store.state.stimulation.hovered_pulse).toStrictEqual({
+      idx: 1,
+      indices: [9, 10],
+      color: "#ffff1",
+    });
+
+    await wrapper.vm.on_pulse_leave();
+    expect(store.state.stimulation.hovered_pulse).toStrictEqual({
+      idx: null,
+      indices: [],
+      color: null,
+    });
+  });
+
+  test("When a user selects 'Duplicate' in  waveform modal, Then the current pulse settings will be added into the pulse order right after selected pulse", async () => {
+    wrapper = mount(StimulationStudioDragAndDropPanel, {
+      store,
+      localVue,
+    });
+
+    await wrapper.setData({ protocol_order: JSON.parse(JSON.stringify(test_protocol_order)) });
+    expect(wrapper.vm.protocol_order).toHaveLength(4);
+    await wrapper.vm.open_modal_for_edit("Monophasic", 3);
+
+    const duplicate_button = wrapper.findAll(".span__button_label").at(1);
+    await duplicate_button.trigger("click");
+
+    expect(wrapper.vm.protocol_order).toHaveLength(5);
   });
 
   test("When a selects the Stimulate Until Complete option in the protocol editor, Then the time unit dropdown should become disabled", async () => {
@@ -407,7 +466,7 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
     expect(wrapper.vm.disable_dropdown).toBe(false);
   });
 
-  test("When a user shift+clicks a delay block to edit duration, Then the new value should be saved upon close", async () => {
+  test("When a user double clicks a delay block to edit duration, Then the new value should be saved upon close", async () => {
     const wrapper = mount(StimulationStudioDragAndDropPanel, {
       store,
       localVue,
@@ -424,7 +483,7 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
     };
 
     const stim_settings = {
-      repeat_delay_interval: {
+      delay_interval: {
         duration: 0,
         unit: "milliseconds",
       },
@@ -434,9 +493,9 @@ describe("StimulationStudioDragAndDropPanel.vue", () => {
       },
     };
 
-    wrapper.vm.protocol_order = test_protocol_order;
+    await wrapper.setData({ protocol_order: test_protocol_order });
     await wrapper.vm.open_modal_for_edit("Delay", idx);
-    expect(wrapper.vm.repeat_delay_modal).toBe("Delay");
+    expect(wrapper.vm.open_delay_modal).toBe(true);
 
     await wrapper.vm.on_modal_close("Save", delay_settings, stim_settings);
     expect(wrapper.vm.protocol_order[idx].pulse_settings.phase_one_duration).toBe(