+++ tests/unit/components/stimulation/StimulationStudioProtocolViewer.spec.js
@@ -2,6 +2,7 @@ import { mount, createLocalVue } from "@vue/test-utils";
 import StimulationStudioProtocolViewer from "@/components/stimulation/StimulationStudioProtocolViewer.vue";
 import StimulationStudioWaveform from "@/components/stimulation/StimulationStudioWaveform.vue";
 import Vuex from "vuex";
+import { convert_x_y_arrays_to_d3_array } from "@/js_utils/waveform_data_formatter.js";
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
@@ -12,25 +13,21 @@ const test_protocol_order = [
   {
     type: "Biphasic",
     src: "placeholder",
-    nest_protocols: [],
     repeat: {
-      number_of_repeats: 0,
-      color: "fffff",
+      number_of_repeats: 1,
+      color: "#ffff1",
     },
     pulse_settings: {
-      phase_one_duration: 300,
-      phase_one_charge: 2,
-      interphase_interval: 500,
-      phase_two_duration: 100,
+      phase_one_duration: 30,
+      phase_one_charge: 20,
+      interphase_interval: 5,
+      phase_two_duration: 10,
       phase_two_charge: -5,
     },
     stim_settings: {
-      repeat_delay_interval: {
-        duration: 0,
-        unit: "milliseconds",
-      },
+      repeat_delay_interval: 10,
       total_active_duration: {
-        duration: 20,
+        duration: 2000,
         unit: "milliseconds",
       },
     },
@@ -38,49 +35,21 @@ const test_protocol_order = [
   {
     type: "Monophasic",
     src: "placeholder",
-    nested_protocols: [
-      {
-        type: "Monophasic",
-        src: "placeholder",
-        nested_protocols: [],
-        repeat: {
-          number_of_repeats: 0,
-          color: "fffff",
-        },
-        pulse_settings: {
-          phase_one_duration: 100,
-          phase_one_charge: -2,
-        },
-      },
-      {
-        type: "Monophasic",
-        src: "placeholder",
-        nested_protocols: [],
-        repeat: {
-          number_of_repeats: 0,
-          color: "fffff",
-        },
-        pulse_settings: {
-          phase_one_duration: 300,
-          phase_one_charge: 6,
-        },
-      },
-    ],
     repeat: {
-      number_of_repeats: 0,
-      color: "fffff",
+      number_of_repeats: 2,
+      color: "#ffff2",
     },
     pulse_settings: {
-      phase_one_duration: 300,
+      phase_one_duration: 30,
       phase_one_charge: 2,
+      interphase_interval: 0,
+      phase_two_duration: 0,
+      phase_two_charge: 0,
     },
     stim_settings: {
-      repeat_delay_interval: {
-        duration: 0,
-        unit: "milliseconds",
-      },
+      repeat_delay_interval: 20,
       total_active_duration: {
-        duration: 0,
+        duration: 1000,
         unit: "milliseconds",
       },
     },
@@ -88,25 +57,21 @@ const test_protocol_order = [
   {
     type: "Delay",
     src: "placeholder",
-    nested_protocols: [],
     repeat: {
-      number_of_repeats: 0,
-      color: "fffff",
+      number_of_repeats: 3,
+      color: "#ffff3",
     },
     pulse_settings: {
-      phase_one_duration: 3000,
+      phase_one_duration: 30,
       phase_one_charge: 0,
       interphase_interval: 0,
       phase_two_duration: 0,
       phase_two_charge: 0,
     },
     stim_settings: {
-      repeat_delay_interval: {
-        duration: "",
-        unit: "milliseconds",
-      },
+      repeat_delay_interval: 20,
       total_active_duration: {
-        duration: 3,
+        duration: 1300,
         unit: "seconds",
       },
     },
@@ -114,22 +79,21 @@ const test_protocol_order = [
   {
     type: "Monophasic",
     src: "placeholder",
-    nested_protocols: [],
     repeat: {
-      number_of_repeats: 0,
-      color: "fffff",
+      number_of_repeats: 2,
+      color: "#ffff4",
     },
     pulse_settings: {
-      phase_one_duration: 300,
-      phase_one_charge: 2,
+      phase_one_duration: 30,
+      phase_one_charge: 50,
+      interphase_interval: 0,
+      phase_two_duration: 0,
+      phase_two_charge: 0,
     },
     stim_settings: {
-      repeat_delay_interval: {
-        duration: 0,
-        unit: "milliseconds",
-      },
+      repeat_delay_interval: 10,
       total_active_duration: {
-        duration: 20,
+        duration: 2000,
         unit: "milliseconds",
       },
     },
@@ -144,6 +108,7 @@ describe("StimulationStudioProtocolViewer.vue", () => {
   beforeEach(async () => {
     store = await NuxtStore.createStore();
   });
+  afterEach(() => jest.clearAllMocks());
 
   test("When exiting instance, Then instance is effectively destroyed", async () => {
     const destroyed_spy = jest.spyOn(StimulationStudioProtocolViewer, "beforeDestroy");
@@ -271,49 +236,90 @@ describe("StimulationStudioProtocolViewer.vue", () => {
     expect(wrapper.vm.repeat_colors).toBe(store.state.stimulation.repeat_colors);
   });
 
-  test("When a user the protocol, Then all datapoints should be deleted", async () => {
-    const wrapper = mount(StimulationStudioWaveform, {
-      store,
-      localVue,
-    });
-
-    const render_spy = jest.spyOn(wrapper.vm, "render_plot");
-    wrapper.vm.$options.watch.data_points.call(wrapper.vm, [
-      [1, 2],
-      [2, 3],
-    ]);
-    wrapper.vm.$options.watch.x_axis_sample_length.call(wrapper.vm, 1000);
-    wrapper.vm.$options.watch.y_min.call(wrapper.vm, 0);
-    wrapper.vm.$options.watch.y_max.call(wrapper.vm, 100);
-
-    expect(render_spy).toHaveBeenCalledTimes(4);
-  });
-
-  test("When a user zooms in or zooms out of the x axis in waveform graph, Then the new graph width will be reflected and the number of the ticks of the axis will change accordingly", async () => {
-    const wrapper = mount(StimulationStudioWaveform, {
-      store,
-      localVue,
-    });
-
-    expect(wrapper.vm.frequency_of_x_ticks).toBe(10);
-    expect(wrapper.vm.div__waveform_graph__dynamic_style).toStrictEqual({ width: "1280px" });
-    await wrapper.setProps({ plot_area_pixel_width: 1800 });
-    expect(wrapper.vm.frequency_of_x_ticks).toBe(15);
-    expect(wrapper.vm.div__waveform_graph__dynamic_style).toStrictEqual({ width: "1880px" });
-  });
-
   test("When a user adds a delay repeat to the end of the protocol, Then it will mutation to state and will automatically update in the waveform graph", async () => {
     const wrapper = mount(StimulationStudioProtocolViewer, {
       store,
       localVue,
     });
 
-    const expected_delay_values = [[3040, 3045]];
     const test_value = 5;
 
     await store.dispatch("stimulation/handle_protocol_order", test_protocol_order);
     await store.dispatch("stimulation/handle_new_rest_duration", test_value);
+
     expect(wrapper.vm.delay_blocks).toBe(store.state.stimulation.delay_blocks);
-    expect(wrapper.vm.delay_blocks).toStrictEqual(expected_delay_values);
+    expect(wrapper.vm.delay_blocks).toStrictEqual([[1305000, 1305005]]);
+  });
+  describe("StimulationStudioWaveform.vue", () => {
+    test("When a user the protocol, Then all datapoints should be deleted", async () => {
+      const wrapper = mount(StimulationStudioWaveform, {
+        store,
+        localVue,
+      });
+
+      const render_spy = jest.spyOn(wrapper.vm, "render_plot");
+      wrapper.vm.$options.watch.data_points.call(wrapper.vm, [
+        [1, 2],
+        [2, 3],
+      ]);
+      wrapper.vm.$options.watch.x_axis_sample_length.call(wrapper.vm, 1000);
+      wrapper.vm.$options.watch.y_min.call(wrapper.vm, 0);
+      wrapper.vm.$options.watch.y_max.call(wrapper.vm, 100);
+
+      expect(render_spy).toHaveBeenCalledTimes(4);
+    });
+
+    test("When a user zooms in or zooms out of the x axis in waveform graph, Then the new graph width will be reflected and the number of the ticks of the axis will change accordingly", async () => {
+      const wrapper = mount(StimulationStudioWaveform, {
+        store,
+        localVue,
+      });
+
+      expect(wrapper.vm.frequency_of_x_ticks).toBe(10);
+      expect(wrapper.vm.div__waveform_graph__dynamic_style).toStrictEqual({ width: "1280px" });
+      await wrapper.setProps({ plot_area_pixel_width: 1800 });
+      expect(wrapper.vm.frequency_of_x_ticks).toBe(15);
+      expect(wrapper.vm.div__waveform_graph__dynamic_style).toStrictEqual({ width: "1880px" });
+    });
+
+    test("When a user hovers over a waveoform tile, Then the corresponding section of the graph will be filled with light opacity", async () => {
+      const wrapper = mount(StimulationStudioWaveform, {
+        store,
+        localVue,
+      });
+
+      await store.dispatch("stimulation/handle_protocol_order", test_protocol_order);
+      const { x_axis_values, y_axis_values } = store.state.stimulation;
+      const d_points = await convert_x_y_arrays_to_d3_array(x_axis_values, y_axis_values);
+      await wrapper.setProps({ data_points: d_points });
+
+      await store.dispatch("stimulation/on_pulse_mouseenter", 1);
+
+      const highlight_line_node = wrapper.find("#highlight_line_node");
+      const highlight_line_path = highlight_line_node.findAll("path");
+
+      expect(highlight_line_path).toHaveLength(1);
+      expect(highlight_line_path.at(0).attributes().opacity).toBe(".15");
+      expect(highlight_line_path.at(0).attributes().fill).toBe("#ffff2");
+    });
+
+    test("When a user leaves hover over a waveoform tile, Then the corresponding section of the graph will no longer be filled", async () => {
+      const wrapper = mount(StimulationStudioWaveform, {
+        store,
+        localVue,
+      });
+
+      await store.dispatch("stimulation/handle_protocol_order", test_protocol_order);
+      const { x_axis_values, y_axis_values } = store.state.stimulation;
+      const d_points = await convert_x_y_arrays_to_d3_array(x_axis_values, y_axis_values);
+      await wrapper.setProps({ data_points: d_points });
+
+      await store.commit("stimulation/on_pulse_mouseleave");
+
+      const highlight_line_node = wrapper.find("#highlight_line_node");
+      const highlight_line_path = highlight_line_node.findAll("path");
+
+      expect(highlight_line_path).toHaveLength(0);
+    });
   });
 });