+++ CHANGELOG.rst
@@ -1,6 +1,39 @@
 Changelog for Mantarray Frontend Components
 ===========================================
 
+1.0.3 (Unreleased)
+------------------
+
+Added:
+^^^^^^
+
+- Additional tooltip message on configuration check icon that notifies user that check cannot be run while live view or a recording is active
+- Added tooltip to barcode input and icon that notifies user it's disabled when live view is active.
+- Added ability to copy pulses in stim studio by double clicking waveform tile and then selecting 'Duplicate'.
+- Added colored background to stim studio waveform graph that corresponds to the waveform tile when hovered over.
+
+Changed
+^^^^^^^
+
+- Frequency in pulse settings modal can now be positive non-integers instead of only positive integers
+- Removed repeat feature in stimulation studio in place of duplicate pulse feature
+- Renamed StimulationStudioRepeatDelayModal to StimulationStudioDelayModal
+
+Fixed
+^^^^^
+
+- Issue with waveform tile overflow in stimulation studio with long protocols
+
+1.0.2 (2022-07-12)
+------------------
+
+Fixed
+^^^^^
+
+- Issue with final protocol markers not being displayed correctly when stopping stim
+- Issue with scanned Beta 2 barcodes coming from the BE being considered invalid
+
+
 1.0.1 (2022-07-11)
 ------------------